refactor(utils): move to src
