+++ packages/web/utils/package.json
@@ -1,9 +1,41 @@
 {
   "name": "@element-plus/utils",
-  "private": true,
+  "version": "3.0.0-alpha.0",
+  "description": "Utils for Element Plus.",
+  "keywords": [
+    "element-plus",
+    "element",
+    "component library",
+    "ui framework",
+    "ui",
+    "utils"
+  ],
+  "homepage": "https://element-plus.org/",
+  "bugs": {
+    "url": "https://github.com/element-plus/element-plus/issues"
+  },
   "license": "MIT",
-  "main": "index.ts",
-  "peerDependencies": {
-    "vue": "^3.2.0"
+  "main": "./src/index.ts",
+  "module": "./src/index.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/element-plus/element-plus.git"
+  },
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "build": "ep-cli build",
+    "clean": "rimraf ./dist"
+  },
+  "dependencies": {
+    "@vue/shared": "^3.2.31",
+    "@vueuse/shared": "^9.1.1",
+    "escape-string-regexp": "^5.0.0",
+    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
+    "lodash-unified": "^1.0.2"
+  },
+  "devDependencies": {
+    "@types/lodash-es": "^4.17.6",
+    "type-fest": "^2.12.2"
   }
 }
