+++ packages/web/utils/src/browser.ts
