+++ packages/web/utils/src/error.ts
