+++ packages/web/utils/tests/raf.test.ts
@@ -2,7 +2,7 @@
 let isClientMocked = false
 
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import { cAF, rAF } from '..'
+import { cAF, rAF } from '../src'
 
 vi.mock('@vueuse/core', () => ({
   get isClient() {
